added files not to large via upload
The 'execution' folder and 'demo-videos' folder are to large to add to the project and so was the 'solution' folder. To add the solution folder I removed the executables from its structure allowing it to be uploaded. There are likely issues running this in your IDE given it relies on those executables to be in the solution folder for things to be accessed by the program appropriately. I am not adding the other folders and they are available to those interested upon request.